feat: add support for DeleteObjects api
